Still Doesn't Work.
Cheated w/ the solution, but keeps barfing up: 0.chunk.js:31837 Warning: Can't call setState on a component that is not yet mounted. This is a no-op, but it might indicate a bug in your application. Instead, assign to `this.state` directly or define a `state = {};` class property with the desired state in the App component.
    in App (at src/index.js:9)
    in StrictMode (at src/index.js:8)